refactor: 바텀 시트에서 화면 이동 로직 개선
