+++ app/src/main/java/woowacourse/movie/view/home/HomeContract.kt
@@ -2,7 +2,6 @@ package woowacourse.movie.view.home
 
 import woowacourse.movie.domain.Movie
 import woowacourse.movie.domain.MovieListItem
-import woowacourse.movie.domain.Showings
 
 interface HomeContract {
     interface Presenter {
@@ -13,10 +12,5 @@ interface HomeContract {
         fun showMovies(movieListItems: List<MovieListItem>)
 
         fun showTheaterSelectDialog(movie: Movie)
-
-        fun navigateToReservation(
-            movie: Movie,
-            showings: Showings,
-        )
     }
 }
