+++ app/src/main/java/woowacourse/movie/view/home/theater/TheaterAdapter.kt
